Renamed all the terminal files
